Quase funcionando

+++ DnsServer.py
@@ -10,51 +10,59 @@ def handle(client, BD, cursor):
     
     while True:
         message = client.recv(1024).decode('utf-8')
+        print(message)
         message = message.split(",")
-        
-        tipo  = message[0]
-        chave = message[1]
-        ip    = message[2]
+        print(message)
+        if len(message)==3:
+            tipo  = message[0]
+            chave = message[1]
+            ip    = message[2]
 
-        if tipo == "Sender":
-            # salva o ip e sua chave primária
-            query = "INSERT INTO Sender (chave, ip) VALUES (%s, %s)"
-            values = (chave, ip)
+            if tipo == "Sender":
+                # salva o ip e sua chave primária
+                query = "INSERT INTO Sender (chave, ip) VALUES (%s, %s)"
+                values = (chave, ip)
 
-            cursor.execute(query, values)
-            BD.commit()
-            
-        elif tipo == "Reciver":
-            try:
-                # Resgata o IP do sender
-                cursor.execute("SELECT ip FROM Sender WHERE chave = %s", (chave,))
-                result = cursor.fetchone()
-
-                sender_ip = result[0]
-
+                cursor.execute(query, values)
+                cursor.fetchone()
+                BD.commit()
+                
+            elif tipo == "Reciver":
                 try:
-                    #Testa para ver se ele está online
-                    socket.create_connection((sender_ip, 5301))
-                    
-                    #Devolve o IP
-                    response = f"IP do sender: {sender_ip}"
-                    client.send(response.encode('utf-8'))
+                    # Resgata o IP do sender
+                    cursor.execute("SELECT ip FROM Sender WHERE chave = %s", (chave,))
+                    result = cursor.fetchone()
 
-                except  ConnectionError:
-                    print("Sender offline será removido")
+                    sender_ip = result[0]
 
-                    query = "DELETE FROM Sender WHERE chave = %s"
-                    cursor.execute(query, (chave,))
+                    try:
+                        #Testa para ver se ele está online
+                        socket.create_connection((sender_ip, 5300))
+                        
+                        #Devolve o IP
+                        response = f"IP do sender: {sender_ip}"
+                        client.send(response.encode('utf-8'))
 
-                    BD.commit()
+                    except  ConnectionError:
+                        print("Sender offline será removido")
+
+                        query = "DELETE FROM Sender WHERE chave = %s"
+                        cursor.execute(query, (chave,))
+
+                        BD.commit()
+                        quit()
+                
+                except:
+                    print("Sender não exite")
+                    client.send("Sender não existe".encode('utf-8'))
                     quit()
-            
-            except:
-                print("Sender não exite")
-                quit()
+
+            else:
+                print("Tipo de pedido não existe")
+
         else:
             print("Recebemos lixo")
-            continue
+            quit()
 
 def tratamento_cliente(client):
     handle(client, BD, cursor)
@@ -71,7 +79,7 @@ BD = mysql.connector.connect(
 
 cursor = BD.cursor()
 cursor.execute("CREATE TABLE IF NOT EXISTS Sender (id INT AUTO_INCREMENT PRIMARY KEY, chave VARCHAR(255), ip VARCHAR(255))")
-
+cursor.fetchone()
 # Testa se a conexão foi bem sucedida.
 
 if BD.is_connected():
